requirements.txt
streamlit
pandas
plotly
gspread
oauth2client
google-generativeai